Code improvement: Addition of InvalidQuantityException for Create Order action

+++ Code/OnLineStore/OnLineStore.Core/BusinessLayer/Contracts/ISalesService.cs
@@ -20,14 +20,14 @@ namespace OnLineStore.Core.BusinessLayer.Contracts
 
         Task<IPagedResponse<OrderInfo>> GetOrdersAsync(int pageSize = 10, int pageNumber = 1, short? currencyID = null, int? customerID = null, int? employeeID = null, short? orderStatusID = null, Guid? paymentMethodID = null, int? shipperID = null);
 
-        Task<ISingleResponse<Order>> GetOrderAsync(Int64 id);
+        Task<ISingleResponse<Order>> GetOrderAsync(long id);
 
         Task<ISingleResponse<CreateOrderRequest>> GetCreateOrderRequestAsync();
 
         Task<ISingleResponse<Order>> CreateOrderAsync(Order header, OrderDetail[] details);
 
-        Task<ISingleResponse<Order>> CloneOrderAsync(int id);
+        Task<ISingleResponse<Order>> CloneOrderAsync(long id);
 
-        Task<IResponse> RemoveOrderAsync(int id);
+        Task<IResponse> RemoveOrderAsync(long id);
     }
 }
